more typmod work along with the rest of the tokenizers... I think
